requirements/base.txt
fastapi==0.63.0
uvicorn==0.13.3
pytest==6.2.2
requests==2.25.1
SQLAlchemy==1.3.23
python-decouple==3.4
